Fix led in Generic

+++ source/computing/math/utils_math.h
@@ -35,67 +35,67 @@ extern "C" {
 #define ABS(n) (((n) < 0) ? -(n) : (n))
 #define AVERAGE_2(a, b) (((a) + (b)) / 2.0)
 
-double calc_percent_s(size_t numerator, size_t denominator);
+float calc_percent_s(size_t numerator, size_t denominator);
 uint32_t calc_percent_uu(uint32_t numerator, uint32_t denominator);
 
-double rad_2_deg(double  rad);
+float rad_2_deg(float  rad);
 bool is_prime(uint32_t number);
 bool is_power_of_two(uint32_t const val);
 
 
-double math_sign(const double val);
+float math_sign(const float val);
 float math_sign_f(const float val);
 
 
-double lo_calc_cos_sample(double time_s,
-		double frequency,
-		double phase_rad,
-		double des_amplitude, double in_offset);
+float lo_calc_cos_sample(float time_s,
+		float frequency,
+		float phase_rad,
+		float des_amplitude, float in_offset);
 
 
-double lo_calc_sin_sample(double time_s,
-		double frequency,
-		double phase_rad,
-		double des_amplitude, double in_offset);
+float lo_calc_sin_sample(float time_s,
+		float frequency,
+		float phase_rad,
+		float des_amplitude, float in_offset);
 
 
-double calc_chirp_sample(uint64_t time_us,
-                         double f_max,
-                         double f_0,
-                         double phase_rad,
-                         double amplitude,
-                         double signal_diration_s);
+float calc_chirp_sample(uint64_t time_us,
+                         float f_max,
+                         float f_0,
+                         float phase_rad,
+                         float amplitude,
+                         float signal_diration_s);
 
-double math_calc_sin_sample(double time_s, double frequency, double phase_ms, double des_amplitude, double in_offset);
+float math_calc_sin_sample(float time_s, float frequency, float phase_ms, float des_amplitude, float in_offset);
 float rad_to_deg_f(float rad);
-double deg_to_rad(double deg);
+float deg_to_rad(float deg);
 bool check_continuity(uint64_t new_val, uint64_t* out_prev_val);
 bool spot_change(uint8_t val);
-double calc_sin_sample(uint64_t time_us, double frequency, double phase_ms, double amplitude, double in_offset);
+float calc_sin_sample(uint64_t time_us, float frequency, float phase_ms, float amplitude, float in_offset);
 #ifdef HAS_AOA
-double aoa_calc_deg(double phase, double wavelength, double dist);
+float aoa_calc_deg(float phase, float wavelength, float dist);
 #endif /*HAS_AOA*/
-double calc_fence_sample(uint64_t time_us, double in_frequency, double in_phase_ms, double amplitude, double in_offset);
-double calc_pwm_sample(uint64_t time_us, double freq, double cur_phase_ms, double des_amplitude, double duty_cycle,
-                       double offset);
+float calc_fence_sample(uint64_t time_us, float in_frequency, float in_phase_ms, float amplitude, float in_offset);
+float calc_pwm_sample(uint64_t time_us, float freq, float cur_phase_ms, float des_amplitude, float duty_cycle,
+                       float offset);
 uint8_t calc_pwm_sample_num(uint64_t time_us, uint32_t period_ms, uint8_t duty, int32_t phase_ms);
 uint32_t int_pow(uint32_t base, uint32_t exp);
 uint64_t ipow(uint32_t base, uint32_t exponenta);
 int8_t two_complement_to_decimal(uint8_t in_code, int significant_bits);
-double calc_saw_sample(uint64_t time_us, double in_frequency, double in_phase_ms, double des_amplitude,
-                       double in_offset);
+float calc_saw_sample(uint64_t time_us, float in_frequency, float in_phase_ms, float des_amplitude,
+                       float in_offset);
 
 
 uint32_t sum_calc_u8(uint8_t* arr, uint32_t size); /*TODO move to array*/
-uint32_t freq2period_ms(double freq_hz);
+uint32_t freq2period_ms(float freq_hz);
 uint8_t toggle_val(uint8_t val);
 
-double calc_dtmf_sample(uint64_t time_us,
-		                double frequency1,
-		                double frequency2,
-		                double phase_ms,
-		                double des_amplitude,
-                        double in_offset);
+float calc_dtmf_sample(uint64_t time_us,
+		                float frequency1,
+		                float frequency2,
+		                float phase_ms,
+		                float des_amplitude,
+                        float in_offset);
 
 #ifdef __cplusplus
 }
